+++ HISTORY.md
@@ -1,6 +1,12 @@
 FORTPY: Revision History
 ======
 
+Revision 1.0.8
+------
+
+- Fixed bug in `fortpy.pysave` that made 2d saving not produce tabular input.
+- Fixed bug that crashed the program when `<isense>` was not present in the `config.xml` file.
+
 Revision 1.0.7
 ------